+++ src/commonMain/kotlin/io/github/koalaplot/core/gestures/TransformGesturesHandler.kt
@@ -3,6 +3,7 @@ package io.github.koalaplot.core.gestures
 import androidx.compose.ui.geometry.Offset
 import androidx.compose.ui.input.pointer.PointerInputScope
 import androidx.compose.ui.unit.IntSize
+import io.github.koalaplot.core.util.ZoomFactor
 
 /**
  * Interface for handling touch-based transformation gestures, such as zoom and pan
@@ -11,9 +12,8 @@ internal interface TransformGesturesHandler {
 
     suspend fun detectTransformGestures(
         scope: PointerInputScope,
-        panLock: Boolean,
-        zoomLock: Boolean,
-        onZoomChange: (size: IntSize, centroid: Offset, zoomX: Float, zoomY: Float) -> Unit,
-        onPanChange: (size: IntSize, pan: Offset) -> Unit,
+        gestureConfig: GestureConfig,
+        onZoomChange: (size: IntSize, centroid: Offset, zoom: ZoomFactor) -> Unit,
+        onPanChange: (size: IntSize, pan: Offset) -> Boolean,
     )
 }