+++ src/commonMain/kotlin/io/github/koalaplot/core/xygraph/XYGraph.kt
@@ -29,6 +29,7 @@ import androidx.compose.ui.unit.IntOffset
 import androidx.compose.ui.unit.IntRect
 import androidx.compose.ui.unit.IntSize
 import androidx.compose.ui.unit.dp
+import io.github.koalaplot.core.gestures.GestureConfig
 import io.github.koalaplot.core.style.KoalaPlotTheme
 import io.github.koalaplot.core.style.LineStyle
 import io.github.koalaplot.core.util.Deg2Rad
@@ -60,37 +61,7 @@ import kotlin.math.sin
  * @param yAxisStyle Style for the y-axis
  * @param yAxisLabels Composable to display labels for specific y-axis values
  * @param yAxisTitle Title for the y-axis
- * @param panEnabled True if the plot can be panned, false to disable. Enabling panning may
- * interfere with scrolling a parent container if the drag point is on the plot.
- * @param zoomEnabled True if the plot can be zoomed, false to disable. Enabling zooming may
- * interfere with scrolling a parent container if the drag point is on the plot.
- * @param allowIndependentZoom True if the zoom can be either only on the X axis, or only on the Y axis,
- * or independently on the X and Y axes at the same time (the behavior depends on the target platform),
- * False if the total zoom factor must be used.
- *
- * Behavior for Android and iOS:
- * True does not mean getting independent zoom coefficients simultaneously for each axis,
- * if the zoom was initiated:
- * - horizontally - the zoom coefficient will change only for the X axis,
- * - vertically - the zoom coefficient will change only for the Y axis.
- *
- * Behavior for Desktop platforms: (EXPERIMENTAL!)
- * True means getting independent zoom coefficients simultaneously or separately for each axis,
- * if the zoom was initiated:
- * - horizontally - the zoom coefficient will change only for the X axis,
- * - vertically - the zoom coefficient will change only for the Y axis,
- * - diagonally - the zoom coefficient will change along the axes X and Y at the same time
- *
- * Behavior for JS and wasmJS: (EXPERIMENTAL!)
- * True means getting independent zoom coefficients simultaneously or separately for each axis,
- * if the zoom was initiated:
- * - horizontally - the zoom coefficient will change only for the X axis,
- * - vertically - the zoom coefficient will change only for the Y axis,
- * - diagonally - the zoom coefficient will change along the axes X and Y at the same time.
- *
- * JS and wasmJS have slight differences in response behavior (for example, zoom coefficients for the same gesture
- * will be interpreted with a difference of several tenths or hundredths), and zoom handling with the mouse wheel
- * scroll while pressing Ctrl/Cmd is not supported (a problem with browser scaling)
+ * @param gestureConfig Configuration for gesture handling. See [GestureConfig]
  * @param content The content to be displayed, which should include one plot for each series to be
  * plotted on this XYGraph.
  */
@@ -110,9 +81,7 @@ public fun <X, Y> XYGraph(
     horizontalMinorGridLineStyle: LineStyle? = KoalaPlotTheme.axis.minorGridlineStyle,
     verticalMajorGridLineStyle: LineStyle? = KoalaPlotTheme.axis.majorGridlineStyle,
     verticalMinorGridLineStyle: LineStyle? = KoalaPlotTheme.axis.minorGridlineStyle,
-    panEnabled: Boolean = false,
-    zoomEnabled: Boolean = false,
-    allowIndependentZoom: Boolean = false,
+    gestureConfig: GestureConfig = GestureConfig(),
     content: @Composable XYGraphScope<X, Y>.() -> Unit
 ) {
     HoverableElementArea(modifier = modifier) {
@@ -147,22 +116,25 @@ public fun <X, Y> XYGraph(
             val yAxisMeasurable = subcompose("yaxis") { Axis(yAxis) }[0]
 
             val chartMeasurable = subcompose("chart") {
-                val panZoomModifier = if (panEnabled || zoomEnabled) {
+                val panZoomModifier = if (gestureConfig.gesturesEnabled) {
                     Modifier.onGestureInput(
                         key1 = xAxisModel,
                         key2 = yAxisModel,
-                        panLock = !panEnabled,
-                        zoomLock = !zoomEnabled,
-                        lockZoomRatio = !allowIndependentZoom,
-                        onZoomChange = { size, centroid, zoomX, zoomY ->
+                        gestureConfig = gestureConfig,
+                        onZoomChange = { size, centroid, zoom ->
                             val normalizedCentroid = centroid.normalizeCentroid(size)
-                            zoomAxis(xAxisModel, size.width, normalizedCentroid.x, zoomX)
-                            zoomAxis(yAxisModel, size.height, normalizedCentroid.y, zoomY)
+                            zoomAxis(xAxisModel, size.width, normalizedCentroid.x, zoom.x)
+                            zoomAxis(yAxisModel, size.height, normalizedCentroid.y, zoom.y)
                         },
                         onPanChange = { size, pan ->
                             val normalizedPan = pan.normalizePan()
-                            panAxis(xAxisModel, size.width, normalizedPan.x)
-                            panAxis(yAxisModel, size.height, normalizedPan.y)
+                            val xPanChanged = panAxis(xAxisModel, size.width, normalizedPan.x)
+                            val yPanChanged = panAxis(yAxisModel, size.height, normalizedPan.y)
+
+                            val allowXPanConsumption = xPanChanged && gestureConfig.panXConsumptionEnabled
+                            val allowYPanConsumption = yPanChanged && gestureConfig.panYConsumptionEnabled
+
+                            return@onGestureInput allowXPanConsumption || allowYPanConsumption
                         }
                     )
                 } else {
@@ -230,8 +202,8 @@ private fun <T> panAxis(
     axis: AxisModel<T>,
     length: Int,
     pan: Float
-) {
-    axis.pan(pan / length.toFloat())
+): Boolean {
+    return axis.pan(pan / length.toFloat())
 }
 
 /**
@@ -664,15 +636,7 @@ private fun DrawScope.drawGridLine(gridLineStyle: LineStyle?, start: Offset, end
  * @param yAxisStyle Style for the y-axis
  * @param yAxisLabels String factory of y-axis label Strings
  * @param yAxisTitle Title for the y-axis
- * @param panEnabled True if the plot can be panned, false to disable. Enabling panning may
- * interfere with scrolling a parent container if the drag point is on the plot.
- * @param zoomEnabled True if the plot can be zoomed, false to disable. Enabling zooming may
- * interfere with scrolling a parent container if the drag point is on the plot.
- * @param allowIndependentZoom `true` if the zoom can be either X-axis only or Y-axis only,
- * `false` if the total zoom factor must be used.
- * `true` does not mean that independent zoom coefficients are obtained simultaneously for each axis,
- * if the zoom was initiated horizontally - the zoom coefficient will change only for the X axis,
- * if the zoom was initiated vertically - the zoom coefficient will change only for the Y axis.
+ * @param gestureConfig Configuration for gesture handling. See [GestureConfig]
  * @param content The content to be displayed within this graph, which should include one plot for each
  * data series to be plotted.
  */
@@ -692,9 +656,7 @@ public fun <X, Y> XYGraph(
     horizontalMinorGridLineStyle: LineStyle? = KoalaPlotTheme.axis.minorGridlineStyle,
     verticalMajorGridLineStyle: LineStyle? = KoalaPlotTheme.axis.majorGridlineStyle,
     verticalMinorGridLineStyle: LineStyle? = KoalaPlotTheme.axis.minorGridlineStyle,
-    panEnabled: Boolean = false,
-    zoomEnabled: Boolean = false,
-    allowIndependentZoom: Boolean = false,
+    gestureConfig: GestureConfig = GestureConfig(),
     content: @Composable XYGraphScope<X, Y>.() -> Unit
 ) {
     XYGraph(
@@ -747,9 +709,7 @@ public fun <X, Y> XYGraph(
         horizontalMinorGridLineStyle,
         verticalMajorGridLineStyle,
         verticalMinorGridLineStyle,
-        panEnabled,
-        zoomEnabled,
-        allowIndependentZoom,
+        gestureConfig,
         content
     )
 }