+++ src/jsMain/kotlin/io/github/koalaplot/core/util/ModifierExtensions.js.kt
@@ -7,11 +7,13 @@ import androidx.compose.ui.input.pointer.PointerEventType
 import androidx.compose.ui.input.pointer.isShiftPressed
 import androidx.compose.ui.input.pointer.onPointerEvent
 import androidx.compose.ui.unit.IntSize
+import io.github.koalaplot.core.gestures.GestureConfig
+import io.github.koalaplot.core.gestures.applyPanLocks
+import io.github.koalaplot.core.gestures.applyZoomLocks
 import io.github.koalaplot.core.gestures.getMaxZoomDeviation
 
 private const val ScrollDeltaMin = 0.75f
 private const val ScrollDeltaMax = 1.25f
-private const val DefaultScrollDelta = 1f
 private const val JsScrollFactor = 100.0f
 
 /**
@@ -23,42 +25,46 @@ private const val JsScrollFactor = 100.0f
 internal actual fun Modifier.onGestureInput(
     key1: Any?,
     key2: Any?,
-    panLock: Boolean,
-    zoomLock: Boolean,
-    lockZoomRatio: Boolean,
-    onZoomChange: (size: IntSize, centroid: Offset, zoomX: Float, zoomY: Float) -> Unit,
-    onPanChange: (size: IntSize, pan: Offset) -> Unit,
+    gestureConfig: GestureConfig,
+    onZoomChange: (size: IntSize, centroid: Offset, zoom: ZoomFactor) -> Unit,
+    onPanChange: (size: IntSize, pan: Offset) -> Boolean,
 ): Modifier = this then Modifier
     .onPointerEvent(PointerEventType.Move) { event ->
-        if (panLock) return@onPointerEvent
+        if (!gestureConfig.panEnabled) return@onPointerEvent
         val change = event.changes.lastOrNull() ?: return@onPointerEvent
-        val result = change.position - change.previousPosition
-        if (result == Offset.Zero) return@onPointerEvent
-        onPanChange(size, result)
-        change.consume()
+
+        val rawPan = change.position - change.previousPosition
+        val pan = rawPan.applyPanLocks(gestureConfig.panXEnabled, gestureConfig.panYEnabled)
+
+        if (pan == Offset.Zero) return@onPointerEvent
+        if (onPanChange(size, pan)) change.consume()
     }
     .onPointerEvent(PointerEventType.Scroll) { event ->
+        if (!gestureConfig.zoomEnabled) return@onPointerEvent
         val change = event.changes.lastOrNull() ?: return@onPointerEvent
-        if (!zoomLock) {
-            val (scrollX, scrollY) = change.scrollDelta
-            val (zoomX, zoomY) = if (lockZoomRatio) {
-                val maxZoomDeviation = getMaxZoomDeviation(
-                    normalizeScrollDeltaToZoom(scrollX),
-                    normalizeScrollDeltaToZoom(scrollY),
-                )
-                maxZoomDeviation to maxZoomDeviation
-            } else if (event.keyboardModifiers.isShiftPressed) {
-                val maxZoomDeviation = getMaxZoomDeviation(
-                    normalizeScrollDeltaToZoom(scrollX),
-                    normalizeScrollDeltaToZoom(scrollY),
-                )
-                maxZoomDeviation to ZoomFactor.NeutralPoint
-            } else {
-                normalizeScrollDeltaToZoom(scrollX) to normalizeScrollDeltaToZoom(scrollY)
-            }
-            onZoomChange(size, change.position, zoomX, zoomY)
-            change.consume()
+
+        val (scrollX, scrollY) = change.scrollDelta
+        val rawZoom = if (!gestureConfig.independentZoomEnabled) {
+            val maxZoomDeviation = getMaxZoomDeviation(
+                normalizeScrollDeltaToZoom(scrollX),
+                normalizeScrollDeltaToZoom(scrollY),
+            )
+            ZoomFactor(maxZoomDeviation, maxZoomDeviation)
+        } else if (event.keyboardModifiers.isShiftPressed) {
+            val maxZoomDeviation = getMaxZoomDeviation(
+                normalizeScrollDeltaToZoom(scrollX),
+                normalizeScrollDeltaToZoom(scrollY),
+            )
+            ZoomFactor(maxZoomDeviation, ZoomFactor.NeutralPoint)
+        } else {
+            ZoomFactor(normalizeScrollDeltaToZoom(scrollX), normalizeScrollDeltaToZoom(scrollY))
         }
+
+        val zoom = rawZoom.applyZoomLocks(gestureConfig.zoomXEnabled, gestureConfig.zoomYEnabled)
+
+        if (zoom == ZoomFactor.Neutral) return@onPointerEvent
+        onZoomChange(size, change.position, zoom)
+        change.consume()
     }
 
 /**
@@ -71,5 +77,5 @@ internal actual fun Modifier.onGestureInput(
  * @param value Scroll Value
  */
 internal fun normalizeScrollDeltaToZoom(value: Float): Float {
-    return (DefaultScrollDelta - value / JsScrollFactor).coerceIn(ScrollDeltaMin..ScrollDeltaMax)
+    return (ZoomFactor.NeutralPoint - value / JsScrollFactor).coerceIn(ScrollDeltaMin..ScrollDeltaMax)
 }