+++ CHANGELOG.md
@@ -10,6 +10,18 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ### Added
 
 - Support for StartAnimationUseCase in bar-chart (GroupedVerticalBarPlot)
+- Ability to disable the consumption of gesture events
+
+### Changed
+
+- Move the logic of enabling/disabling pan and zoom to the gesture handler
+- A separate object has been created for the gesture configuration GestureConfig
+- Removing the gesture logic "pastTouchSlop", in practice it turned out to be inconsistent when capturing panning
+  from the parent container
+
+### Fixed
+
+- Discrete panning in X and Y axes (#104)
 
 ## [0.8.0]